+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.9 NextJS demo, better watch mode
+
+- A tidy GraphQL example
+- A fix for watching config file on linux systems
+- New `browser` option in the config file to indicate which browser should be launched.
+- Cache globs now match hostname + pathname, not just pathname
+- Logfiles are now one per day instead a single giant file
+- Fix unit tests, unskip disabled tests
+
 ## 0.0.8 Bugfixes, config watchmode
 
 - Bugfix: `trust` list